+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - chore: remove volume when stop container
 - chore: remove option user on .envcontainer.yaml
 - chore: add new flag to start with vscode
+- fix: regex to validate ports pattern
 
 ### Fixed