+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - chore: remove volume when stop container
 - chore: remove option user on .envcontainer.yaml
+- chore: add new flag to start with vscode
 - feat: add new flag to start with vscode
 - feat: add commant to list envcontainers 
 - fix: regex to validate ports pattern